yolo-detection/requirements.txt
paho-mqtt
tflite-runtime
opencv-python
